+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+ - FusionTelemetryMunger supports registering multiple mapping sets under existing mapping types
+
 ## [5.0.0] - 2019-05-16
 ### Added
  - Added attributes for decorating telemetry properties, methods and events